requirements.txt
# Core NLP & Embeddings
sentence-transformers>=2.2.2
spacy>=3.7.0
nltk>=3.8.1
transformers>=4.35.0

# Knowledge Graph & Graph Processing
networkx>=3.2.1
python-louvain>=0.16
leidenalg>=0.10.1
igraph>=0.11.3

# Vector Storage & Similarity
chromadb>=0.4.22
faiss-cpu>=1.7.4
scikit-learn>=1.3.2

# LLM Integration
ollama>=0.1.6
langchain>=0.1.0
langchain-community>=0.0.10

# PDF Processing
pypdf>=3.17.4
pdfplumber>=0.10.3

# Utilities
pyyaml>=6.0.1
tqdm>=4.66.1
numpy>=1.24.0
pandas>=2.1.4

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# Optional: Evaluation
# ragas>=0.1.0
